Lots of refactoring
Pads are now drawn over the map as sprites. This necessitated a whole new collision system. Also fixed the issue where the ship would look like it was in a valid place but exploded anyway. Pads change colors. Mail carried by the mailman and the exhaust from the ship are now overlain on the icons. Flips are used to save on icon space. Ship(s) are now an object container, though the pad search code still, for now, requires there to be only one ship and uses the "ship" global for it. Added the blue pad color for the "post office" pad. Concept of used (green) and unused (red) pads enforced.