Added the HUD to show where unvisited pads are.
I'm working on HUD stuff. I've been through a couple designs; the one I have now is the one I liked best. Short dashes point to the remaining pads, even if they are on screen. I was going to make a full HUD around the ship but it looked terrible. Anyway. Next is to make the victory condition activate after all pads are visited.