Level end detection, some Pluto reskinning
We now detect the end of the level and point the player toward the exit portal. I'm working on reskinning but it's not going super well. Right now I am trying to make it into a game ASAP, and then I can work on refinement. A LOT of refinement. But I am trying to get this to a point where I can post it as a WIP to the PICO 8 BBS.